requirements.txt
# Core Dependencies
streamlit>=1.28.1
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.1.3

# UI and Visualization
plotly>=5.17.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Development and Testing (optional)
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
